program/crud/demo05/requirements.txt
# 基础框架
Flask==3.1.0
Werkzeug==3.1.3

# 数据库
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.39

# REST API
flask-restful==0.3.10

# 安全认证
Flask-JWT-Extended==4.7.1
passlib==1.7.4  # 用于密码哈希

# 接口限流
Flask-Limiter==3.12

# API文档
apispec==6.8.1
flask-smorest==0.45.0  # OpenAPI支持

# 数据校验
marshmallow==3.26.1

# 配置管理
python-dotenv==1.0.1

# 单元测试
pytest==8.3.5
pytest-cov==4.1.0

# 附加工具
watchdog==4.0.0  # 开发时文件监控
